requirements.txt
altair==5.2.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
duckdb==0.9.2
extra-streamlit-components==0.1.60
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gspread==5.12.4
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.2.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==1.5.3
pillow==10.2.0
plotly==5.18.0
protobuf==4.25.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sql_metadata==2.10.0
sqlparse==0.4.4
st-gsheets-connection==0.0.3
streamlit==1.30.0
streamlit-authenticator==0.2.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
